Cleaned up code and changed location of s3 credentials
