+++ config/dbInit.js
@@ -3,16 +3,20 @@ const mysql = require('mysql2/promise');
 require('dotenv').config();
 
 async function dbInit() {
-  const connection = await mysql.createConnection({
-    host: process.env.DB_HOST,
-    port: process.env.DB_PORT,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-  });
+  if (process.env.JAWSDB_URL) {
+    return;
+  } else {
+    const connection = await mysql.createConnection({
+      host: process.env.DB_HOST,
+      port: process.env.DB_PORT,
+      user: process.env.DB_USER,
+      password: process.env.DB_PASSWORD,
+    });
 
-  await connection.query(
-    `CREATE DATABASE IF NOT EXISTS ${process.env.DB_NAME}`
-  );
+    await connection.query(
+      `CREATE DATABASE IF NOT EXISTS ${process.env.DB_NAME}`
+    );
+  }
 }
 
 module.exports = dbInit;
