+++ heroku.env.EXAMPLE
@@ -1,0 +1,6 @@
+# Cookie for session cookie encryption
+COOKIE_SECRET='' # recommended to use 32 random characters
+
+# Database config credentials
+S3_ACCESS_KEY_ID=''
+S3_SECRET_ACCESS_KEY=''
