+++ server.js
@@ -1,58 +1,34 @@
 const path = require('path');
 const express = require('express');
-const session = require('express-session');
-const expressHandlebars = require('express-handlebars');
-const dbInit = require('./config/dbInit');
-const SequelizeStore = require('connect-session-sequelize')(session.Store);
-const multer = require('multer');
-
 const routes = require('./controllers');
+const expressSession = require('express-session');
+const expressSessionInfo = require('./config/session');
+const expressHandlebars = require('express-handlebars');
+const SequelizeStore = require('connect-session-sequelize')(session.Store);
+const dbInit = require('./config/dbInit');
 
+// initialize the database
 dbInit();
 
+// import sequelize connection
 const sequelize = require('./config/connection');
 
 const app = express();
 
-const sessionInfo = {
-  secret: process.env.COOKIE_SECRET,
-  cookie: {
-    maxAge: 86400000, //24 hours in milliseconds
-    httpOnly: true,
-    secure: false,
-    sameSite: 'strict',
-  },
-  resave: false,
-  saveUninitialized: true,
-  store: new SequelizeStore({
-    db: sequelize,
-  }),
-};
-
-const uploads = multer({ dest: __dirname + './public/game-images' });
-
-
-
-// var storage = multer.diskStorage({
-//   destination: function (req, file, callback) {
-//     callback(null, __dirname + './public/game-images');
-//   },
-//   filename: function (req, file, callback) {
-//     callback(null, file.originalname);
-//   },
-// });
-
-app.use(session(sessionInfo));
-
+// set up handlebars.js engine
 app.engine('handlebars', expressHandlebars());
 app.set('view engine', 'handlebars');
 
+// set up sessions
+app.use(expressSession(expressSessionInfo));
+
+// set up routes
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(express.static(path.join(__dirname, 'public')));
-
 app.use(routes);
 
+// turn on connection to db and server
 sequelize.sync({ force: false }).then(() => {
-  app.listen(process.env.APP_PORT, () => console.log('Now listening'));
+  app.listen(process.env.PORT, () => console.log('Now listening'));
 });
